inference_server/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
pillow>=10.0.0
python-multipart==0.0.6
numpy>=1.24.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
pydantic[email]>=1.10.0
